Add files via upload
Added the basic JSON parse function. We can adjust it later to gather more information, however for now it returns a list of two variables, the first being the string which the Google API assigned to it, the second being a floating point number of the percent likelihood the image was about that label. You will have to do a function call to the function I made in that file in order to get the list output. We can adjust it easily later for further more complicated use if need be.