Added a Folder for Unused Code
The folder here contains unused code because our final implementation for the class did not include everything that we envisioned at first. Although the code we used for the class is simpler, it is incompatible with these files. If we want to include different ways to analyze the code, then we can use this in the future to observe different things, such as if there are landmarks in the picture and how does that impact the image and the likes the image gets, among other things. Ideally we would have implemented this, however due to time constraints we are instead leaving it here. Most of the files work; just not in tandem. The big issue is that the database does not work with these current files.